pyproject.toml
[tool.poetry]
name = "seq2seq-attention-pytorch-lightning"
version = "0.1.0"
description = "Pytorch-Lightning Seq2Seq with attention tutorial"
authors = ["ksopyla <krzysztofsopyla@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pytorch-lightning = "^1.2.5"
torch = "^1.8.0"
pytorch-nlp = {git = "https://github.com/PetrochukM/PyTorch-NLP.git"}

[tool.poetry.dev-dependencies]
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
